+++ stac_fastapi/pgstac/tests/api/test_api.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timedelta
+import pytest
 
 STAC_CORE_ROUTES = [
     "GET /",
@@ -281,3 +282,73 @@ async def test_search_line_string_intersects(
     assert resp.status_code == 200
     resp_json = resp.json()
     assert len(resp_json["features"]) == 1
+
+
+@pytest.mark.asyncio
+async def test_search_forwarded_header(
+    load_test_data, app_client, load_test_collection
+):
+    coll = load_test_collection
+    item = load_test_data("test_item.json")
+    await app_client.post(f"/collections/{coll.id}/items", json=item)
+    resp = await app_client.post(
+        "/search",
+        json={
+            "collections": [item["collection"]],
+        },
+        headers={"Forwarded": "proto=https;host=test:1234"},
+    )
+    features = resp.json()["features"]
+    assert len(features) > 0
+    for feature in features:
+        for link in feature["links"]:
+            assert link["href"].startswith("https://test:1234/")
+
+
+@pytest.mark.asyncio
+async def test_search_x_forwarded_headers(
+    load_test_data, app_client, load_test_collection
+):
+    coll = load_test_collection
+    item = load_test_data("test_item.json")
+    await app_client.post(f"/collections/{coll.id}/items", json=item)
+    resp = await app_client.post(
+        "/search",
+        json={
+            "collections": [item["collection"]],
+        },
+        headers={
+            "X-Forwarded-Proto": "https",
+            "X-Forwarded-Port": "1234",
+        },
+    )
+    features = resp.json()["features"]
+    assert len(features) > 0
+    for feature in features:
+        for link in feature["links"]:
+            assert link["href"].startswith("https://test:1234/")
+
+
+@pytest.mark.asyncio
+async def test_search_duplicate_forward_headers(
+    load_test_data, app_client, load_test_collection
+):
+    coll = load_test_collection
+    item = load_test_data("test_item.json")
+    await app_client.post(f"/collections/{coll.id}/items", json=item)
+    resp = await app_client.post(
+        "/search",
+        json={
+            "collections": [item["collection"]],
+        },
+        headers={
+            "Forwarded": "proto=https;host=test:1234",
+            "X-Forwarded-Proto": "http",
+            "X-Forwarded-Port": "4321",
+        },
+    )
+    features = resp.json()["features"]
+    assert len(features) > 0
+    for feature in features:
+        for link in feature["links"]:
+            assert link["href"].startswith("https://test:1234/")